+++ graph.go
@@ -165,7 +165,7 @@ type (
 	// "color" with value "red".
 	EdgeProperties struct {
 		Attributes map[string]string
-		Weight     int
+		Weight     float64
 		Data       any
 	}
 
@@ -200,7 +200,7 @@ func IntHash(v int) int {
 // EdgeWeight returns a function that sets the weight of an edge to the given
 // weight. This is a functional option for the [graph.Graph.Edge] and
 // [graph.Graph.AddEdge] methods.
-func EdgeWeight(weight int) func(*EdgeProperties) {
+func EdgeWeight(weight float64) func(*EdgeProperties) {
 	return func(e *EdgeProperties) {
 		e.Weight = weight
 	}
@@ -269,13 +269,13 @@ func EdgeCopy[K comparable](e Edge[K]) (K, K, func(*EdgeProperties)) {
 // "color" with value "red".
 type VertexProperties struct {
 	Attributes map[string]string
-	Weight     int
+	Weight     float64
 }
 
 // VertexWeight returns a function that sets the weight of a vertex to the given
 // weight. This is a functional option for the [graph.Graph.Vertex] and
 // [graph.Graph.AddVertex] methods.
-func VertexWeight(weight int) func(*VertexProperties) {
+func VertexWeight(weight float64) func(*VertexProperties) {
 	return func(e *VertexProperties) {
 		e.Weight = weight
 	}
